+++ Runtime/Affise.cs
@@ -18,7 +18,6 @@ using AffiseAttributionLib.SKAd;
 using AffiseAttributionLib.Utils;
 #if (UNITY_ANDROID || UNITY_IOS) && !UNITY_EDITOR
 using AffiseAttributionLib.Native;
-using AffiseAttributionLib.Executors;
 #endif
 
 namespace AffiseAttributionLib
@@ -86,17 +85,7 @@ namespace AffiseAttributionLib
         public static void SendEventNow(AffiseEvent affiseEvent, OnSendSuccessCallback success, OnSendFailedCallback failed)
         {
 #if (UNITY_ANDROID || UNITY_IOS) && !UNITY_EDITOR
-            var contextExecutor = new ContextThreadExecutor();
-            _native?.SendEventNow(affiseEvent, () =>
-            {
-                contextExecutor.Run(success.Invoke);
-            }, (status) =>
-            {
-                contextExecutor.Run(() =>
-                {
-                    failed.Invoke(status);
-                });
-            });
+            _native?.SendEventNow(affiseEvent, success, failed);
 #else
             _api?.ImmediateSendToServerUseCase.SendNow(affiseEvent, success, failed);
 #endif
@@ -120,14 +109,7 @@ namespace AffiseAttributionLib
         public static void RegisterDeeplinkCallback(DeeplinkCallback callback)
         {
 #if (UNITY_ANDROID || UNITY_IOS) && !UNITY_EDITOR
-            var contextExecutor = new ContextThreadExecutor();
-            _native?.RegisterDeeplinkCallback((value) =>
-            {
-                contextExecutor.Run(() =>
-                {
-                    callback.Invoke(value);
-                });
-            });
+            _native?.RegisterDeeplinkCallback(callback);
 #else
             _api?.DeeplinkManager.SetDeeplinkCallback(callback);
 #endif
@@ -298,14 +280,7 @@ namespace AffiseAttributionLib
         public static void GetReferrerUrl(ReferrerCallback callback)
         {
 #if (UNITY_ANDROID || UNITY_IOS) && !UNITY_EDITOR
-            var contextExecutor = new ContextThreadExecutor();
-            _native?.GetReferrerUrl((value) =>
-            {
-                contextExecutor.Run(() =>
-                {
-                    callback.Invoke(value);
-                });
-            });
+            _native?.GetReferrerUrl(callback);
 #else
             callback.Invoke(NotSupported);
 #endif
@@ -317,14 +292,7 @@ namespace AffiseAttributionLib
         public static void GetReferrerUrlValue(ReferrerKey key, ReferrerCallback callback)
         {
 #if (UNITY_ANDROID || UNITY_IOS) && !UNITY_EDITOR
-            var contextExecutor = new ContextThreadExecutor();
-            _native?.GetReferrerUrlValue(key, (value) =>
-            {
-                contextExecutor.Run(() =>
-                {
-                    callback.Invoke(value);
-                });
-            });
+            _native?.GetReferrerUrlValue(key, callback);
 #else
             callback.Invoke(NotSupported);
 #endif
@@ -386,14 +354,7 @@ namespace AffiseAttributionLib
             public static void RegisterAppForAdNetworkAttribution(ErrorCallback completionHandler)
             {
 #if (UNITY_IOS) && !UNITY_EDITOR
-                var contextExecutor = new ContextThreadExecutor();
-                _native?.RegisterAppForAdNetworkAttribution((error) =>
-                {
-                    contextExecutor.Run(() =>
-                    {
-                        completionHandler.Invoke(error);
-                    });
-                });
+                _native?.RegisterAppForAdNetworkAttribution(completionHandler);
 #else
                 completionHandler.Invoke(NotSupported);
 #endif
@@ -405,14 +366,7 @@ namespace AffiseAttributionLib
             public static void UpdatePostbackConversionValue(int fineValue, CoarseValue coarseValue, ErrorCallback completionHandler)
             {
 #if (UNITY_IOS) && !UNITY_EDITOR
-                var contextExecutor = new ContextThreadExecutor();
-                _native?.UpdatePostbackConversionValue(fineValue, coarseValue, (error) =>
-                {
-                    contextExecutor.Run(() =>
-                    {
-                        completionHandler.Invoke(error);
-                    });
-                });
+                _native?.UpdatePostbackConversionValue(fineValue, coarseValue, completionHandler);
 #else
                 completionHandler.Invoke(NotSupported);
 #endif
@@ -424,14 +378,7 @@ namespace AffiseAttributionLib
             public static void GetReferrerOnServer(ReferrerCallback callback)
             {
 #if (UNITY_ANDROID || UNITY_IOS) && !UNITY_EDITOR
-                var contextExecutor = new ContextThreadExecutor();
-                _native?.GetReferrerOnServer((result) =>
-                {
-                    contextExecutor.Run(() =>
-                    {
-                        callback.Invoke(result);
-                    });
-                });
+                _native?.GetReferrerOnServer(callback);
 #else
                 callback.Invoke(NotSupported);
 #endif
@@ -443,14 +390,7 @@ namespace AffiseAttributionLib
             public static void GetReferrerOnServerValue(ReferrerKey key, ReferrerCallback callback)
             {
 #if (UNITY_ANDROID || UNITY_IOS) && !UNITY_EDITOR
-                var contextExecutor = new ContextThreadExecutor();
-                _native?.GetReferrerOnServerValue(key, (result) =>
-                {
-                    contextExecutor.Run(() =>
-                    {
-                        callback.Invoke(result);
-                    });
-                });
+                _native?.GetReferrerOnServerValue(key, callback);
 #else
                 callback.Invoke(NotSupported);
 #endif
@@ -465,14 +405,7 @@ namespace AffiseAttributionLib
             public static void GetStatus(AffiseModules module, OnKeyValueCallback onComplete)
             {
 #if (UNITY_ANDROID || UNITY_IOS) && !UNITY_EDITOR
-                var contextExecutor = new ContextThreadExecutor();
-                _native?.GetStatus(module, (result) =>
-                {
-                    contextExecutor.Run(() =>
-                    {
-                        onComplete.Invoke(result);
-                    });
-                });
+                _native?.GetStatus(module, onComplete);
 #else
                 _api?.ModuleManager.Status(module, onComplete);
 #endif
@@ -508,14 +441,7 @@ namespace AffiseAttributionLib
             public static void LinkResolve(string uri, AffiseLinkCallback callback)
             {
 #if (UNITY_ANDROID || UNITY_IOS) && !UNITY_EDITOR
-                var contextExecutor = new ContextThreadExecutor();
-                _native?.LinkResolve(uri, (result) =>
-                {
-                    contextExecutor.Run(() =>
-                    {
-                        callback.Invoke(result);
-                    });
-                });
+                _native?.LinkResolve(uri, callback);
 #else
                 Api<IAffiseLinkApi>(AffiseModules.Link)?.LinkResolve(uri, callback);
 #endif
@@ -527,14 +453,7 @@ namespace AffiseAttributionLib
             public static void FetchProducts(List<string> ids, AffiseResultCallback<AffiseProductsResult> callback)
             {
 #if (UNITY_ANDROID || UNITY_IOS) && !UNITY_EDITOR
-                var contextExecutor = new ContextThreadExecutor();
-                _native?.FetchProducts(ids, (result) =>
-                {
-                    contextExecutor.Run(() =>
-                    {
-                        callback.Invoke(result);
-                    });
-                });           
+                _native?.FetchProducts(ids, callback);           
 #else
                 UnityEngine.Debug.LogWarning($"{NotSupported} - FetchProducts");
                 callback.Invoke(AffiseResult<AffiseProductsResult>.Failure(NotSupported));
@@ -551,14 +470,7 @@ namespace AffiseAttributionLib
             )
             {
 #if (UNITY_ANDROID || UNITY_IOS) && !UNITY_EDITOR
-                var contextExecutor = new ContextThreadExecutor();
-                _native?.Purchase(product, type, (result) =>
-                {
-                    contextExecutor.Run(() =>
-                    {
-                        callback.Invoke(result);
-                    });
-                });
+                _native?.Purchase(product, type, callback);
 #else               
                 UnityEngine.Debug.LogWarning($"{NotSupported} - Purchase");
                 callback.Invoke(AffiseResult<AffisePurchasedInfo>.Failure(NotSupported));
@@ -586,14 +498,7 @@ namespace AffiseAttributionLib
             public static void Validate(DebugOnValidateCallback callback)
             {
 #if (UNITY_ANDROID || UNITY_IOS) && !UNITY_EDITOR
-                var contextExecutor = new ContextThreadExecutor();
-                _native?.Validate((status) =>
-                {
-                    contextExecutor.Run(() =>
-                    {
-                        callback.Invoke(status);
-                    });
-                });
+                _native?.Validate(callback);
 #else
                 _api?.DebugValidateUseCase?.Validate(callback);
 #endif
@@ -607,14 +512,7 @@ namespace AffiseAttributionLib
             public static void Network(DebugOnNetworkCallback callback)
             { 
 #if (UNITY_ANDROID || UNITY_IOS) && !UNITY_EDITOR
-                var contextExecutor = new ContextThreadExecutor();
-                _native?.Network((request, response) =>
-                {
-                    contextExecutor.Run(() =>
-                    {
-                        callback.Invoke(request, response);
-                    });
-                });
+                _native?.Network(callback);
 #else
                 _api?.DebugNetworkUseCase.OnRequest(callback);
 #endif